Adding query for searching cv by company

+++ src/main/java/com/example/demo/controller/api/TrainingRestController.java
@@ -53,7 +53,7 @@ public class TrainingRestController {
   }
 
   @PutMapping("edit/{id}")
-  public ResponseEntity<Object> put(@PathVariable Integer id, @RequestBody Training trainingEdit) {
+  public ResponseEntity<Object> put(@PathVariable Long id, @RequestBody Training trainingEdit) {
     Training training = trainingRepository.findById(id).get();
     if (training.getCvPerson().getId() == trainingEdit.getCvPerson().getId()) {
       training.setName(trainingEdit.getName());
@@ -69,7 +69,7 @@ public class TrainingRestController {
   }
 
   @DeleteMapping("delete/{id}")
-  public ResponseEntity<Object> delete(@PathVariable Integer id, @RequestParam("cvId") Integer cvId) {
+  public ResponseEntity<Object> delete(@PathVariable Long id, @RequestParam("cvId") Integer cvId) {
     if (trainingRepository.countByCVId(cvId) == 1) {
       CVPerson cvPerson = cvPersonRepository.findById(cvId).get();
       cvPerson.setPercentage_progress(cvPerson.getPercentage_progress() - 20);
